applications: connectivity_bridge: Update ring_buf buffer datatype
Update ring_buf datatype to match updated ring_buf api.

+++ applications/connectivity_bridge/src/modules/uart_handler.c
@@ -60,7 +60,7 @@ struct uart_rx_buf {
 
 struct uart_tx_buf {
 	struct ring_buf rb;
-	uint32_t buf[UART_BUF_SIZE];
+	uint8_t buf[UART_BUF_SIZE];
 };
 
 BUILD_ASSERT((sizeof(struct uart_rx_buf) % UART_SLAB_ALIGNMENT) == 0);
